Adjust tramp-file-name-regexp to work with looking-at
Recent updates to tramp have introduced an ending match of empty
string that does not work together with looking-at as used in
hpath.

A promising alternative to this change would be to use file-remote-p
in combination with a regexp that finds a possible filename after
point. This would remove the dependency on the internal regexp in
tramp.