Try and avoid using `hyperb:stack-frame`
Scanning the stack frame can be useful in some unusual cases, but
should be avoided when a straightforward let-binding does the trick.

* hmouse-drv.el (hkey--within-help): New var.
(hkey-help): Bind it.
* hibtypes.el (pathname): Consult it.
(hibtypes--within-org-link-outside-org-mode): New var.
(org-link-outside-org-mode): Use it to detect recursion.

* hui-mouse.el (hyp--within-smart-org): New var.
(smart-org): Bind it.
(hkey-alist): Consult it.
(-flatten): Move declaration to where we know the function is
actually available.