Improve tag/symbol finding, <> and {} handling
hui-mouse.el (hkey-alist): Add clause to handle Smart Key when in
a vertico minibuffer window; make it behave like standard vertico
which exits and selects the first minibuffer line, not the candidate.

If in the minibuffer and reading an argument with vertico run the
vertico command on {M-RET} which accepts the first line of minibuffer
input, rather than any candidate.