hyrolo-normalize-mode-function - Normalize Org/Markdown buf display
hui:link-possible-types - When a single-line region is active in the
referent buffer, create a link-to-string-match link to the text in
the region.

hpath:line-column-instance-regexp - Add 'I<num>' support for
link-to-string-match links; jumps to instance <num> of match.

ibut:insert-text - Add instance number support and always insert
link-to-string-match using compact hash tag syntax.