Fix window/minibuffer latency issue
This (sit-for) causes an annoying bit of latency every time a new
window or minibuffer is opened. The extra delay makes Emacs feel very
clunky, and almost put me off using Hyperbole more. I'm unsure what
the consequences are to removing this, but I would like to see a
better solution to any bugs this sleep may be hiding.