hpath:shorten - Expand & abbrev, don't shorten gbut activated paths
hmouse-tag.el (smart-prog-tag):
hui-mouse.el (hkey-alist): Change Assist Key behavior when on a programming
symbol/tag.  If the symbol is a variable holding an in-memory Hyperbole button,
then display a help buffer with the button's attributes.  Otherwise, if an Elisp
symbol, then display a help buffer with symbol documentation, if any, to match
the Hyperbole manual documentation.  For any other language symbol, display its
definition, just as the Action Key does.